Refactored namespace Salary into EmployeeSalary
