+++ DomainLayer/Models/Employee/IEmployeeModel.cs
@@ -6,7 +6,7 @@ using DomainLayer.Models.EmployeeEmploymentInfo;
 using DomainLayer.Models.EmployeeFinancialInfo;
 using DomainLayer.Models.EmployeePersonalInfo;
 using DomainLayer.Models.Leave;
-using DomainLayer.Models.Salary;
+using DomainLayer.Models.EmployeeSalary;
 using DomainLayer.Models.User;
 
 namespace DomainLayer.Models.Employee
